Use same capitalization convention for PN as the rest of ldmx-sw
