+++ include/SimCore/PhotoNuclearModels/BertiniNothingHardModel.h
@@ -9,8 +9,8 @@
 #include <G4ProcessManager.hh>
 
 #include "Framework/Configure/Parameters.h"
-#include "SimCore/PhotonuclearModel.h"
-#include "SimCore/PhotonuclearModels/BertiniEventTopologyProcess.h"
+#include "SimCore/PhotoNuclearModel.h"
+#include "SimCore/PhotoNuclearModels/BertiniEventTopologyProcess.h"
 namespace simcore {
 
 class BertiniNothingHardProcess : public BertiniEventTopologyProcess {
@@ -36,11 +36,11 @@ class BertiniNothingHardProcess : public BertiniEventTopologyProcess {
   double Emin_;
 };
 
-class BertiniNothingHardModel : public PhotonuclearModel {
+class BertiniNothingHardModel : public PhotoNuclearModel {
  public:
   BertiniNothingHardModel(const std::string& name,
                           const framework::config::Parameters& parameters)
-      : PhotonuclearModel{name, parameters},
+      : PhotoNuclearModel{name, parameters},
         threshold_{parameters.getParameter<double>("hard_particle_threshold")},
         Zmin_{parameters.getParameter<int>("zmin")},
         Emin_{parameters.getParameter<double>("emin")},
