+++ src/SimCore/PhotoNuclearModels/BertiniAtLeastNProductsModel.cxx
@@ -1,34 +1,33 @@
 
-#include "SimCore/PhotonuclearModels/BertiniSingleNeutronModel.h"
+#include "SimCore/PhotoNuclearModels/BertiniAtLeastNProductsModel.h"
 namespace simcore {
 
-bool BertiniSingleNeutronProcess::acceptEvent() const {
-  int Nhard{0};
-  int Nhard_neutron{0};
+bool BertiniAtLeastNProductsProcess::acceptEvent() const {
   int secondaries{theParticleChange.GetNumberOfSecondaries()};
+  int matchingCount{0};
   for (int i{0}; i < secondaries; ++i) {
     const auto secondary{theParticleChange.GetSecondary(i)->GetParticle()};
     const auto pdgCode{secondary->GetDefinition()->GetPDGEncoding()};
-    if (skipCountingParticle(pdgCode)) {
-      continue;
-    }
     const auto energy{secondary->GetKineticEnergy()};
-    if (energy > threshold_) {
-      Nhard++;
-      if (pdgCode == 2112) {
-        Nhard_neutron++;
+    if (std::find(std::begin(pdg_ids_), std::end(pdg_ids_), pdgCode) !=
+        std::end(pdg_ids_)) {
+      if (energy > threshold_) {
+        ++matchingCount;
       }
     }
+    if (matchingCount >= min_products_) {
+      return true;
+    }
   }
-  return Nhard == 1 && Nhard_neutron == 1;
+  return false;
 }
 
-void BertiniSingleNeutronModel::ConstructGammaProcess(
+void BertiniAtLeastNProductsModel::ConstructGammaProcess(
     G4ProcessManager* processManager) {
   auto photoNuclearProcess{
       new G4HadronInelasticProcess("photonNuclear", G4Gamma::Definition())};
-  auto model{new BertiniSingleNeutronProcess{threshold_, Zmin_, Emin_,
-                                             count_light_ions_}};
+  auto model{new BertiniAtLeastNProductsProcess{threshold_, Zmin_, Emin_,
+                                                pdg_ids_, min_products_}};
   model->SetMaxEnergy(15 * CLHEP::GeV);
   addPNCrossSectionData(photoNuclearProcess);
   photoNuclearProcess->RegisterMe(model);
@@ -36,4 +35,4 @@ void BertiniSingleNeutronModel::ConstructGammaProcess(
 }
 }  // namespace simcore
 
-DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniSingleNeutronModel)
+DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniAtLeastNProductsModel)
