+++ src/SimCore/PhotoNuclearModels/BertiniModel.cxx
@@ -1,0 +1,3 @@
+#include "SimCore/PhotoNuclearModels/BertiniModel.h"
+
+DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniModel)
