+++ src/SimCore/PhotoNuclearModels/BertiniSingleNeutronModel.cxx
@@ -1,7 +1,10 @@
-#include "SimCore/PhotonuclearModels/BertiniNothingHardModel.h"
+
+#include "SimCore/PhotoNuclearModels/BertiniSingleNeutronModel.h"
 namespace simcore {
 
-bool BertiniNothingHardProcess::acceptEvent() const {
+bool BertiniSingleNeutronProcess::acceptEvent() const {
+  int Nhard{0};
+  int Nhard_neutron{0};
   int secondaries{theParticleChange.GetNumberOfSecondaries()};
   for (int i{0}; i < secondaries; ++i) {
     const auto secondary{theParticleChange.GetSecondary(i)->GetParticle()};
@@ -11,18 +14,21 @@ bool BertiniNothingHardProcess::acceptEvent() const {
     }
     const auto energy{secondary->GetKineticEnergy()};
     if (energy > threshold_) {
-      return false;
+      Nhard++;
+      if (pdgCode == 2112) {
+        Nhard_neutron++;
+      }
     }
   }
-  return true;
+  return Nhard == 1 && Nhard_neutron == 1;
 }
 
-void BertiniNothingHardModel::ConstructGammaProcess(
+void BertiniSingleNeutronModel::ConstructGammaProcess(
     G4ProcessManager* processManager) {
   auto photoNuclearProcess{
       new G4HadronInelasticProcess("photonNuclear", G4Gamma::Definition())};
-  auto model{new BertiniNothingHardProcess{threshold_, Zmin_, Emin_,
-                                           count_light_ions_}};
+  auto model{new BertiniSingleNeutronProcess{threshold_, Zmin_, Emin_,
+                                             count_light_ions_}};
   model->SetMaxEnergy(15 * CLHEP::GeV);
   addPNCrossSectionData(photoNuclearProcess);
   photoNuclearProcess->RegisterMe(model);
@@ -30,4 +36,4 @@ void BertiniNothingHardModel::ConstructGammaProcess(
 }
 }  // namespace simcore
 
-DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniNothingHardModel)
+DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniSingleNeutronModel)
