--- src/SimCore/PhotonuclearModels/BertiniModel.cxx
@@ -1,3 +1,0 @@
-#include "SimCore/PhotonuclearModels/BertiniModel.h"
-
-DECLARE_PHOTONUCLEAR_MODEL(simcore::BertiniModel)
